# backend/docs/swagger.yaml
basePath: /
definitions:
  models.AllowedEmailRequest:
    properties:
      email:
        description: Email для добавления/удаления
        example: new.participant@example.com
        type: string
      event_id:
        description: ID мероприятия
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
    type: object
...
  models.CreateEventRequest:
    properties:
      address:
        description: Адрес проведения
        example: Ул. Программистов, 15
        type: string
      description:
        description: Описание мероприятия
        example: Крупнейшее событие года в сфере технологий.
        type: string
      ends_at:
        description: Время окончания мероприятия
        example: "2025-12-01T18:00:00Z"
        type: string
      name:
        description: Название мероприятия
        example: Ежегодная конференция IT
        type: string
      starts_at:
        description: Время начала мероприятия
        example: "2025-12-01T10:00:00Z"
        type: string
    type: object
  models.EventResponse:
    properties:
      address:
        type: string
      creator_email:
        type: string
      description:
        type: string
      ends_at:
        type: string
      id:
        type: string
      name:
        type: string
      starts_at:
        type: string
    type: object
  models.ListEventResponse:
    properties:
      address:
        type: string
      description:
        type: string
      ends_at:
        type: string
      id:
        type: string
      name:
        type: string
      starts_at:
        type: string
    type: object
...
  models.UpdateEventRequest:
    properties:
      address:
        description: Новый адрес
        example: Санкт-Петербург, Экспоцентр
        type: string
      description:
        description: Новое описание мероприятия
        example: Новое описание события.
        type: string
      ends_at:
        description: Новое время окончания
        example: "2025-12-01T19:00:00Z"
        type: string
      event_id:
        description: ID мероприятия для обновления
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      name:
        description: Новое название мероприятия
        example: Обновленное название конференции
        type: string
      starts_at:
        description: Новое время начала
        example: "2025-12-01T10:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис для управления событиями, картами и точками интереса.
  title: Event Designer API
  version: "1.0"
paths:
  /admin/events:
    patch:
      consumes:
      - application/json
      description: Обновляет поля мероприятия. Используется метод PATCH, можно передавать
        только те поля, которые нужно изменить.
      parameters:
      - description: Поля для обновления мероприятия
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Мероприятие успешно обновлено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON / Event ID required / Invalid event ID format
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Database error / Failed to update event
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Обновить информацию о мероприятии
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Создает новое мероприятие. Доступно только администраторам.
      parameters:
      - description: Данные для создания мероприятия
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание, возвращает ID и имя
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON / Name is required
          schema:
            type: string
        "500":
          description: Authentication context missing / Failed to create event
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Создать мероприятие
      tags:
      - Events
  /admin/events/{eventID}:
    delete:
      description: Удаляет мероприятие по его ID. Доступно только администраторам.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: eventID
        required: true
        type: string
      responses:
        "204":
          description: Мероприятие успешно удалено (No Content)
        "400":
          description: Invalid event ID format in URL
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Failed to delete event
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить мероприятие
      tags:
      - Events
...
  /admin/events/allowed-emails:
    delete:
      consumes:
      - application/json
      description: Удаляет email из списка разрешенных участников для данного мероприятия.
      parameters:
      - description: ID мероприятия и Email для удаления
        in: body
        name: allowedEmail
        required: true
        schema:
          $ref: '#/definitions/models.AllowedEmailRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Email успешно удален (No Content)
        "400":
          description: Invalid JSON / Required fields missing / Invalid event ID format
          schema:
            type: string
        "500":
          description: Failed to delete allowed email or email/event not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить разрешенный email
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Добавляет email в список разрешенных участников для данного мероприятия.
      parameters:
      - description: ID мероприятия и Email для добавления
        in: body
        name: allowedEmail
        required: true
        schema:
          $ref: '#/definitions/models.AllowedEmailRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Email успешно добавлен (No Content)
        "400":
          description: Invalid JSON / Required fields missing / Invalid event ID format
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Database error / Failed to add allowed email
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Добавить разрешенный email
      tags:
      - Events
...
  /events:
    get:
      description: Получает список мероприятий. Для администраторов - все созданные
        ими мероприятия и мероприятия на которые они добавлены. Для участников - мероприятия,
        на которые их email разрешен.
      produces:
      - application/json
      responses:
        "200":
          description: Список мероприятий
          schema:
            items:
              $ref: '#/definitions/models.ListEventResponse'
            type: array
        "500":
          description: Authentication context missing / Failed to retrieve events
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить список мероприятий
      tags:
      - Events
  /events/{event_id}:
    get:
      description: Получает детальную информацию о мероприятии. Доступно создателю
        мероприятия и участникам, чьи email разрешены.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Полная информация о мероприятии
          schema:
            $ref: '#/definitions/models.EventResponse'
        "400":
          description: Invalid event ID format
          schema:
            type: string
        "403":
          description: Access denied for this event
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Database error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить мероприятие по ID
      tags:
      - Events
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"